+++ R/recode_to.R
@@ -6,7 +6,7 @@
 #'   Useful when recoding dummy variables with 1/2 values to 0/1 values,  or
 #'   recoding scales from 1-4 to 0-3 etc.
 #'   \code{recode_to_if()} is a scoped variant of \code{recode_to()}, where
-#'   recoding will be applied only to those variable that match the
+#'   recoding will be applied only to those variables that match the
 #'   logical condition of \code{predicate}.
 #'
 #' @seealso \code{\link{rec}} for general recoding of variables and \code{\link{set_na}}