+++ R/split_var.R
@@ -4,7 +4,7 @@
 #' @description Recode numeric variables into equal sized groups, i.e. a
 #'   variable is cut into a smaller number of groups at specific cut points.
 #'   \code{split_var_if()} is a scoped variant of \code{split_var()}, where
-#'   transformation will be applied only to those variable that match the
+#'   transformation will be applied only to those variables that match the
 #'   logical condition of \code{predicate}.
 #'
 #' @seealso \code{\link{group_var}} to group variables into equal ranged groups,